pyproject.toml
[project]
name = "endcli"
version = "0.0.1"
authors = [
    {name="Arvin Vitug", email="arvinvitug@gmail.com"}
    ]
description = "A cli tool to pull information from an ESPN NBA fantasy league" 
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "discord",
    "python-dotenv",
    "espn_api"
]

[build-system]
requires = [
    "setuptools >= 61.0"
]
build-backend = "setuptools.build_meta"

[project.scripts]
endcli = "main:main"